// snl-gms-mock/packages/common-model/src/ts/channel-segment/types.ts
export enum TimeseriesType {
  WAVEFORM = 'WAVEFORM',
  FK_SPECTRA = 'FK_SPECTRA',
  DETECTION_FEATURE_MAP = 'DETECTION_FEATURE_MAP'
}